tmuxビルドのディレクトリ構成を変更

+++ bin/build-tmux.sh
@@ -6,6 +6,14 @@ mkdir -p "$env" && cd "$env" || exit 1
 env="`pwd`"
 src="$env/src/tmux"
 dest="$env/dest/tmux"
+profile_d="$env/profile.d"
+mkdir -p "$profile_d"
+
+requires="automake libevent-devel ncurses-devel"
+if ! rpm -q $requires >/dev/null 2>&1; then
+  echo "yum install $requires"
+  exit 1
+fi
 
 # develop version
 rm -rf "$src"
@@ -14,8 +22,6 @@ cd "$src" || exit 1
 
 # 参考：罫線パッチ http://d.hatena.ne.jp/emonkak/20110521/1305970697
 patch -p1 < "$base/bin/build-tmux.sh.border_patch"
-
-sudo yum install -y automake libevent-devel ncurses-devel || exit 1
 
 #libevent2が必要なので1.xしか無い場合はソースで入れる
 if rpm -q libevent-devel-1.\* >/dev/null 2>&1; then
@@ -26,11 +32,10 @@ if rpm -q libevent-devel-1.\* >/dev/null 2>&1; then
   ) || exit 1
   export CPPFLAGS="-I$env/dest/libevent2/include"
   export LDFLAGS="-L$env/dest/libevent2/lib"
-  echo "export LD_LIBRARY_PATH=\"\$LD_LIBRARY_PATH:$env/dest/libevent2/lib\"" > ~/.profile.d/10-dotfiles-tmux-libevent2.sh
+  echo "export LD_LIBRARY_PATH=\"\$LD_LIBRARY_PATH:$env/dest/libevent2/lib\"" > "$profile_d"/10-dotfiles-tmux-libevent2.sh
 fi
 
 sh autogen.sh || exit 1
 ./configure --prefix="$dest" && make && make install || exit 1
 
-mkdir -p ~/.profile.d || exit 1
-echo "export PATH=\"$dest/bin:\$PATH\"" > ~/.profile.d/10-dotfiles-tmux.sh
+echo "export PATH=\"$dest/bin:\$PATH\"" > "$profile_d"/10-dotfiles-tmux.sh
